requirements.txt
elasticsearch
aiohttp
python-dotenv
google-genai
python-dateutil
pydantic
rich
